+++ packages/cli/src/runners/generate.ts
@@ -2,7 +2,7 @@ import process from 'node:process'
 import { type Config, safeBuild, setup } from '@kubb/core'
 import { createLogger, LogMapper } from '@kubb/core/logger'
 import { Presets, SingleBar } from 'cli-progress'
-import { colors } from 'consola/utils'
+import pc from 'picocolors'
 import type { Args } from '../commands/generate.ts'
 import { executeHooks } from '../utils/executeHooks.ts'
 import { getErrorCauses } from '../utils/getErrorCauses.ts'
@@ -83,7 +83,7 @@ export async function generate({ input, config, progressCache, args }: GenerateP
     logger,
   })
 
-  logger.emit('start', `Building ${logger.logLevel !== LogMapper.silent ? colors.dim(inputPath!) : ''}`)
+  logger.emit('start', `Building ${logger.logLevel !== LogMapper.silent ? pc.dim(inputPath!) : ''}`)
 
   const { files, error } = await safeBuild({
     config: definedConfig,
@@ -109,7 +109,7 @@ export async function generate({ input, config, progressCache, args }: GenerateP
 
   if (error && logger.consola) {
     logger.consola?.resumeLogs()
-    logger.consola.error(`Build failed ${logger.logLevel !== LogMapper.silent ? colors.dim(inputPath!) : ''}`)
+    logger.consola.error(`Build failed ${logger.logLevel !== LogMapper.silent ? pc.dim(inputPath!) : ''}`)
 
     logger.consola.box({
       title: `${config.name || ''}`,
@@ -137,7 +137,7 @@ export async function generate({ input, config, progressCache, args }: GenerateP
     await executeHooks({ hooks: config.hooks, logger })
   }
 
-  logger.consola?.log(`⚡Build completed ${logger.logLevel !== LogMapper.silent ? colors.dim(inputPath!) : ''}`)
+  logger.consola?.log(`⚡Build completed ${logger.logLevel !== LogMapper.silent ? pc.dim(inputPath!) : ''}`)
 
   logger.consola?.box({
     title: `${config.name || ''}`,