+++ packages/cli/src/utils/getSummary.ts
@@ -2,7 +2,7 @@ import path from 'node:path'
 
 import { randomCliColour } from '@kubb/core/logger'
 
-import { colors } from 'consola/utils'
+import pc from 'picocolors'
 
 import { parseHrtimeToSeconds } from './parseHrtimeToSeconds.ts'
 
@@ -32,20 +32,20 @@ export function getSummary({ pluginManager, filesCreated, status, hrStart, confi
   const meta = {
     plugins:
       status === 'success'
-        ? `${colors.green(`${buildStartPlugins.length} successful`)}, ${pluginsCount} total`
-        : `${colors.red(`${failedPlugins?.length ?? 1} failed`)}, ${pluginsCount} total`,
+        ? `${pc.green(`${buildStartPlugins.length} successful`)}, ${pluginsCount} total`
+        : `${pc.red(`${failedPlugins?.length ?? 1} failed`)}, ${pluginsCount} total`,
     pluginsFailed: status === 'failed' ? failedPlugins?.map((name) => randomCliColour(name))?.join(', ') : undefined,
     filesCreated: filesCreated,
-    time: `${colors.yellow(`${elapsedSeconds}s`)}`,
+    time: `${pc.yellow(`${elapsedSeconds}s`)}`,
     output: path.isAbsolute(config.root) ? path.resolve(config.root, config.output.path) : config.root,
   } as const
 
   logs.add(
     [
-      [`${colors.bold('Plugins:')}        ${meta.plugins}`, true],
-      [`${colors.dim('Failed:')}          ${meta.pluginsFailed || 'none'}`, !!meta.pluginsFailed],
-      [`${colors.bold('Generated:')}      ${meta.filesCreated} files in ${meta.time}`, true],
-      [`${colors.bold('Output:')}         ${meta.output}`, true],
+      [`${pc.bold('Plugins:')}        ${meta.plugins}`, true],
+      [`${pc.dim('Failed:')}          ${meta.pluginsFailed || 'none'}`, !!meta.pluginsFailed],
+      [`${pc.bold('Generated:')}      ${meta.filesCreated} files in ${meta.time}`, true],
+      [`${pc.bold('Output:')}         ${meta.output}`, true],
     ]
       .map((item) => {
         if (item.at(1)) {