+++ packages/config-tsup/tsup.config.ts
@@ -1,5 +1,0 @@
-import { defineConfig } from 'tsup'
-
-import { optionsCJS, optionsESM } from './src/index.ts'
-
-export default defineConfig([optionsCJS, optionsESM])