+++ packages/plugin-oas/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @kubb/swagger
 
+## 3.16.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @kubb/core@3.16.2
+  - @kubb/oas@3.16.2
+  - @kubb/react@3.16.2
+
 ## 3.16.1
 
 ### Patch Changes